+++ src/shared-components/alert/index.js
@@ -55,7 +55,7 @@ class Alert extends React.Component {
   }
 
   exit = () => {
-    const { onExit, ...rest} = this.props;
+    const { onExit, ...rest } = this.props;
     this.setState({ exiting: true });
 
     // eslint-disable-next-line no-undef
@@ -63,8 +63,8 @@ class Alert extends React.Component {
 
     // eslint-disable-next-line no-undef
     window.setTimeout(() => {
-      onExit({...rest});
-      this.setState({exited: true });
+      onExit({ ...rest });
+      this.setState({ exited: true });
     }, ANIMATION_DELAY);
   };