+++ src/shared-components/index.js
@@ -1,5 +1,6 @@
 export { default as Accordion } from './accordion';
 export { default as Alert } from './alert';
+export { default as Banner } from './banner';
 export { default as Button, RoundButton } from './button';
 export { default as Checkbox } from './checkbox';
 export { default as Chip } from './chip';
@@ -8,5 +9,6 @@ export { default as ImmersiveModal } from './immersiveModal';
 export { default as LoadingSpinner } from './loadingSpinner';
 export { default as Modal } from './modal';
 export { default as OffClickWrapper } from './offClickWrapper';
+export { default as Dropdown } from './dropdown';
 export { default as RadioButton } from './radioButton';
 export { default as Typography, style as TYPOGRAPHY_STYLE } from './typography';